requirements.txt
numpy==1.21.6
packaging==24.2
Pillow==11.1.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.6.2.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
